clinker-core: extract executor format I/O builders into registry/ingest
Move the format-writer construction path out of `executor/mod.rs` into a
new `registry.rs` (alongside the `WriterRegistry` it serves) and the
format-reader / source-ingest path into a new `ingest.rs`:

- `registry.rs` — `WriterRegistry`, `build_format_writer`, the writer
  factory, split-policy + split-naming helpers, and the per-format
  writer-config builders.
- `ingest.rs` — `build_format_reader`, multi-file reader assembly,
  schema-coercion wrapping, the per-source driver thread
  (`ingest_source` / `drive_record_source`), event-time stamping, and
  the per-format reader-config builders.

`ingest_source` and its `IngestTaskOutcome` are widened from file-private
to `pub(super)` so the executor entry point that spawns the ingest
threads reaches them across the new module boundary; `build_format_writer`
keeps its existing `pub(crate)` visibility for the output / correlation
dispatch arms. `WriterRegistry` and the moved `build_format_writer` are
re-exported from `mod.rs`, so `executor::WriterRegistry` /
`executor::build_format_writer` resolve unchanged.

Pure relocation: no symbol renames, no public-API change, every existing
test passes unmodified.

Closes #259
